Fix 'a yee' detection (#57)
Currently it will trigger if it detects 'A' and 'yee' in the post, regardless of where they are. A sentence like "Ong Yee Kuan said something. A new hope." would trigger the detection. This fixes it by ensuring that 'A' & 'yee' must always be beside one another with only non alphabet characters separating them (eg. A.Yee or A_Yee)